Crate linera_explorer

This module provides web files to run a block explorer from Linera service node and Linera indexer.

Structs

Config

Config type dealt with localstorage.

Data

type for Vue data.

GQuery

GraphQL query type (for subscriptions).

Enums

AddressKind

Protocol

Functions

data

Initializes Vue data.

route

save_config

Saves config to local storage.

short_app_id

short_crypto_hash

start

Initializes pages and subscribes to notifications.